Add Spotify and generated dataset
Dataset:
 - Spotify (361 profiles, 30 candidates, 48 ranks)
 - generated (synthetic) dataset (12360 profiles, 30 candidates, 48 ranks)- balanced with regards to Borda Winner.



\paragraph{Representation 1.}
In this representation the set of features is the set of candidates $C$ and the value for each feature is the Borda score of the feature. The profile from Example~\ref{ex:running} is given in Table~\ref{tab:rep1}.

\begin{table}
\centering
\caption{The profile from Example~\ref{ex:running} in Representation 1. }\label{tab:rep1}
\begin{tabular}{c|c|c|c||c}
%\hline
f-a & f-b  & f-c & f-d & label\\
\hline
11 & 12 & 9 & 10 & b  
%1 voter  & $d \succ b \succ a \succ c$ \\
%1 voter  & $d \succ c \succ a \succ b$ \\
%1 voter  & $b \succ d \succ c \succ a$ \\
%1 voter  & $c \succ d \succ b \succ a$ \\
%\hline
\end{tabular}
\end{table}



%This factorised representation transforms original list of ranks into Borda count dataset, such that the feature is candidate and the value of the feature is his total number of Borda points, i.e., depending on the position of the candidate in each rank Borda rule assigns n-1 points to a top ranked choice, n-2 points to second ranked choice, down to 0 points for a bottom ranked alternative. 
The shape of new representation is  $( \# profiles, \# candidates)$. Here, the shape of the Spotify model is: $(361, 30)$ and the generated model: $(12360, 30)$.



\paragraph{Representation 2.}
The features in this representation are the set of candidates together with the position at which each candidate can be ranked. For example, if the set of candidates has four candidates we obtain 16 features, one for each position in which each candidate can be ranked. The value of the feature is the number of times the candidate occurs in the particular position in the profile.  The shape of the new representation is: $$( \# profiles, |C|\cdot|C|).$$ Thus, for given  $N = 30$  candidates  and length of each vote (ranking) equal to $30$ positions, we obtain $900$ features. Here, shape of the Spotify model is: $(361, 900)$ and of the generated model: $(12360, 900)$.

The profile from Example~\ref{ex:running} is given in Table~\ref{tab:rep2}.

\begin{table}[h!]
\centering
\caption{The profile from Example~\ref{ex:running} in Representation 2. }\label{tab:rep2}
\begin{tabular}{c|c|c|c |c|c|c|c| c|c|c|c| c|c|c|c ||c}
%\hline
f-a1 & f-a2 & f-a3 & f-a4 & f-b1 & f-b2 & f-b3 & f-b4 & f-c1 & f-c2 & f-c3 & f-c4 & f-d1 & f-d2 & f-d3 & f-d4 & label\\
\hline
3 & 0 & 2 & 2 & 1 & 4 & 1 & 1 & 1 & 1 & 4 & 1 & 2 & 2 & 0 & 3 & b  
%\hline
\end{tabular}
\end{table}




\paragraph{Representation 3.}
Here, we obtain the features again from the set of candidates, but we now consider one feature to be a unique pair of candidates. The value of feature $ab$ is 1 if $a \succ_1 b$ and 0 otherwise. 
%This representation of data consist of making pairwise comparison of the candidates and scoring by "1" whenever there is agreement and "0" in case of disagreement for each individual preferance order, such that the features are combination of pairs of candidates, and value of the feature becomes "0" or "1" value. We join repeated combination of pairwise comparison to previous profile transformation. Thus,
For  given $N=30$ candidates we have  $870$ possible pairs. For a profile of $48$ voters we obtain 41,760 features.  The shape of the   representation is: $$( \# profiles, \binom{|C|}{2} \cdot |V|).$$ Here, the shape of the Spotify model is: $(361, 41760)$ and the generated model: $(12360, 41760)$.



\paragraph{Representation 4.}
Here, the set of features is the set of unique pairs of candidates. The value of the feature is the number of voters that prefer the first candidate to the second. 
The profile from Example~\ref{ex:running} is given in Table~\ref{tab:rep4}.

\begin{table}[h!]
\centering
\caption{The profile from Example~\ref{ex:running} in Representation 4. }\label{tab:rep4}
\begin{tabular}{c|c|c|c |c|c|c|c| c|c|c|c||c}
%\hline
ab & ac & ad & ba & bc & bd & ca & cb & cd & da & db & dc  & label\\
\hline
4 & 4 & 3 &
3 & 5 & 4 & 
3 & 2 & 5 & 
4 & 2 & 3  & b  
%\hline
\end{tabular}
\end{table}

%Similarly to Representation 3, here we use pairs of candidates to build features, but now the value o the fethis factorisation has been build based on pairwise comparison, such that the features are combination of pairwise comparison of candidates, but this time the feature becomes the sum of points (agreements in preference orders). Instead of joining the dataset of transformed profile, we keep cumulative points for each combination of pairwise comparison. 

For given $N=30$ candidates we have  $870$ features. The shape of the representation is: $$( \# profiles, \binom{|C|}{2}).$$Here, shape of the Spotify model is: $(361, 870)$ and the generated model: $(12360, 870)$.